Unconfirmed minor fixes (#488)
