+++ schema/json_build_config.json
@@ -8,6 +8,8 @@
       "description": "serde support for message",
       "attr": "#[serde_with::serde_as]\n#[derive(serde::Serialize, serde::Deserialize, Eq, std::hash::Hash)]",
       "paths": [
+        "structs.RecentTransactionAndObservationRequest",
+        "structs.TransactionAndObservations",
         "structs.TransportInfo",
         "structs.GenericTypedValue",
         "structs.SchemaDef",
