+++ schema/src/proto/structs.proto
@@ -857,6 +857,16 @@ message RecentDiscoveryTransactionsRequest {
   optional int64 min_time = 2;
 }
 
+message TransactionAndObservations {
+  Transaction transaction = 1;
+  repeated Observation observation = 2;
+}
+
+message RecentTransactionAndObservationRequest {
+  optional int64 limit = 1;
+  optional int64 offset = 3;
+}
+
 message Request {
   optional GossipTransactionRequest gossip_transaction_request = 1;
   optional GossipObservationRequest gossip_observation_request = 2;
@@ -885,6 +895,7 @@ message Request {
   UtxoId utxo_valid_request = 25;
   RecentDiscoveryTransactionsRequest recent_transactions_request = 26;
   Hash lookup_transaction_request = 27;
+  RecentTransactionAndObservationRequest recent_transaction_and_observation_request = 28;
 }
 
 message HealthResponse {
@@ -927,6 +938,7 @@ message Response {
   UtxoValidResponse utxo_valid_response = 17;
   RecentDiscoveryTransactionsResponse recent_discovery_transactions_response = 18;
   Transaction lookup_transaction_response = 19;
+  repeated TransactionAndObservations recent_transaction_and_observation_response = 20;
 }
 
 message QueryTransactionRequest {
