+++ src/hardware/trezor.rs
@@ -6,12 +6,12 @@ use bitcoin::util::psbt::serialize::Serialize;
 use itertools::Itertools;
 use redgold_keys::proof_support::ProofSupport;
 use redgold_keys::TestConstants;
-use redgold_keys::transaction_support::{InputSupport, TransactionBuilderSupport};
+use redgold_keys::transaction_support::InputSupport;
 use redgold_schema::{error_info, ErrorInfoContext, SafeBytesAccess, SafeOption, structs, WithMetadataHashable};
-use redgold_schema::structs::{AddressInfo, ErrorInfo, Hash, Output, Proof, Signature, Transaction, CurrencyAmount, UtxoEntry, Input, UtxoId};
-use redgold_schema::transaction::amount_data;
-use redgold_schema::transaction_builder::TransactionBuilder;
+use redgold_schema::structs::{AddressInfo, CurrencyAmount, ErrorInfo, Hash, Input, Output, Proof, Signature, Transaction, UtxoEntry, UtxoId};
+use crate::core::transact::tx_builder_supports::TransactionBuilder;
 use redgold_keys::util::mnemonic_words::HDPathCursor;
+use crate::core::transact::tx_builder_supports::TransactionBuilderSupport;
 use crate::util::cmd::run_cmd;
 use crate::util::keys::{public_key_from_bytes, ToPublicKeyFromLib};
 use crate::util::init_logger_once;
